chore(EMS-3153): no pdf - split up policy - check/change answers e2e tests
