+++ .github/workflows/test.yml
@@ -202,7 +202,8 @@ jobs:
             "application-submission/multiple-policy-type/**/*.spec.js",
             "application-submission/single-policy-type/**/*.spec.js",
             "cannot-skip-flow/**/*.spec.js",
-            "check-your-answers/policy/**/*.spec.js",
+            "check-your-answers/policy/multiple-policy/**/*.spec.js",
+            "check-your-answers/policy/single-policy/**/*.spec.js",
             "check-your-answers/your-business/**/*.spec.js",
             "check-your-answers/your-buyer/**/*.spec.js",
             "dashboard/**/*.spec.js",
@@ -243,7 +244,11 @@ jobs:
             "policy/broker/**/*.spec.js",
             "policy/broker-details/**/*.spec.js",
             "policy/broker-confirm-address/**/*.spec.js",
-            "policy/change-your-answers/**/*.spec.js",
+            "policy/change-your-answers/broker/**/*.spec.js",
+            "policy/change-your-answers/multiple-policy/**/*.spec.js",
+            "policy/change-your-answers/name-on-policy/**/*.spec.js",
+            "policy/change-your-answers/pre-credit-period/**/*.spec.js",
+            "policy/change-your-answers/single-policy/**/*.spec.js",
             "policy/check-your-answers/**/*.spec.js",
             "policy/different-name-on-policy/**/*.spec.js",
             "policy/loss-payee/**/*.spec.js",
