+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/policy/change-your-answers/multiple-policy/change-your-answers-policy-multiple-policy-type.spec.js
@@ -1,17 +1,18 @@
-import partials from '../../../../../../../partials';
-import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
-import { INSURANCE_FIELD_IDS } from '../../../../../../../constants/field-ids/insurance';
-import { field, status, summaryList } from '../../../../../../../pages/shared';
-import application from '../../../../../../../fixtures/application';
-import formatCurrency from '../../../../../../../helpers/format-currency';
-import { USD } from '../../../../../../../fixtures/currencies';
-import { createTimestampFromNumbers, formatDate } from '../../../../../../../helpers/date';
+import partials from '../../../../../../../../partials';
+import { INSURANCE_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+import { field, summaryList } from '../../../../../../../../pages/shared';
+import { multipleContractPolicyExportValuePage } from '../../../../../../../../pages/insurance/policy';
+import formatCurrency from '../../../../../../../../helpers/format-currency';
+import application from '../../../../../../../../fixtures/application';
+import { USD } from '../../../../../../../../fixtures/currencies';
+import { createTimestampFromNumbers, formatDate } from '../../../../../../../../helpers/date';
 
 const {
   ROOT: INSURANCE_ROOT,
   POLICY: {
-    SINGLE_CONTRACT_POLICY_CHECK_AND_CHANGE,
-    SINGLE_CONTRACT_POLICY_TOTAL_CONTRACT_VALUE_CHECK_AND_CHANGE,
+    MULTIPLE_CONTRACT_POLICY_CHECK_AND_CHANGE,
+    MULTIPLE_CONTRACT_POLICY_EXPORT_VALUE_CHECK_AND_CHANGE: exportValueRoute,
   },
   CHECK_YOUR_ANSWERS: {
     TYPE_OF_POLICY,
@@ -23,7 +24,10 @@ const {
   POLICY: {
     CONTRACT_POLICY: {
       REQUESTED_START_DATE,
-      SINGLE: { CONTRACT_COMPLETION_DATE, TOTAL_CONTRACT_VALUE },
+      MULTIPLE: { TOTAL_MONTHS_OF_COVER },
+    },
+    EXPORT_VALUE: {
+      MULTIPLE: { TOTAL_SALES_TO_BUYER, MAXIMUM_BUYER_WILL_OWE },
     },
   },
 } = INSURANCE_FIELD_IDS;
@@ -32,8 +36,9 @@ const { taskList } = partials.insurancePartials;
 
 const task = taskList.submitApplication.tasks.checkAnswers;
 
-const getFieldVariables = (fieldId, referenceNumber) => ({
-  route: SINGLE_CONTRACT_POLICY_CHECK_AND_CHANGE,
+const getFieldVariables = (fieldId, referenceNumber, route = MULTIPLE_CONTRACT_POLICY_CHECK_AND_CHANGE) => ({
+  route,
+  checkYourAnswersRoute: TYPE_OF_POLICY,
   newValueInput: '',
   fieldId,
   referenceNumber,
@@ -43,14 +48,14 @@ const getFieldVariables = (fieldId, referenceNumber) => ({
 
 const baseUrl = Cypress.config('baseUrl');
 
-context('Insurance - Change your answers - Policy - Single contract policy - Summary List', () => {
+context('Insurance - Change your answers - Policy - Multiple contract policy - Summary List', () => {
   let url;
   let referenceNumber;
 
   before(() => {
     cy.completeSignInAndGoToApplication({}).then(({ referenceNumber: refNumber }) => {
       referenceNumber = refNumber;
-      cy.completePrepareApplicationSinglePolicyType({ referenceNumber });
+      cy.completePrepareApplicationMultiplePolicyType({ referenceNumber });
 
       task.link().click();
 
@@ -68,88 +73,24 @@ context('Insurance - Change your answers - Policy - Single contract policy - Sum
     cy.deleteApplication(referenceNumber);
   });
 
-  describe('single policy type answers', () => {
+  describe('multiple policy type answers', () => {
     describe(REQUESTED_START_DATE, () => {
       const fieldId = REQUESTED_START_DATE;
 
-      const newStartDate = {
+      const fieldVariables = getFieldVariables(fieldId, referenceNumber);
+
+      const newAnswer = {
         ...application.POLICY[fieldId],
         year: application.POLICY[fieldId].year + 1,
       };
 
-      const newEndDate = {
-        ...application.POLICY[fieldId],
-        year: newStartDate.year + 1,
-      };
-
-      const fieldVariables = {
-        startDate: {
-          ...getFieldVariables(fieldId, referenceNumber),
-          newValueInput: newStartDate.year,
-        },
-        endDate: {
-          ...application.POLICY[CONTRACT_COMPLETION_DATE],
-          ...getFieldVariables(CONTRACT_COMPLETION_DATE, referenceNumber),
-          newValueInput: newEndDate.year,
-        },
-      };
-
       describe('when clicking the `change` link', () => {
         beforeEach(() => {
           cy.navigateToUrl(url);
         });
 
-        it(`should redirect to ${SINGLE_CONTRACT_POLICY_CHECK_AND_CHANGE}`, () => {
+        it(`should redirect to ${MULTIPLE_CONTRACT_POLICY_CHECK_AND_CHANGE}`, () => {
           cy.navigateToUrl(url);
-
-          cy.checkChangeLinkUrl(fieldVariables.startDate, referenceNumber);
-        });
-      });
-
-      describe('form submission with a new answer', () => {
-        beforeEach(() => {
-          cy.navigateToUrl(url);
-
-          summaryList.field(fieldId).changeLink().click();
-
-          cy.changeAnswerField(fieldVariables.startDate, field(fieldId).yearInput(), false);
-
-          cy.changeAnswerField(
-            fieldVariables.endDate,
-            field(CONTRACT_COMPLETION_DATE).yearInput(),
-          );
-        });
-
-        it(`should redirect to ${TYPE_OF_POLICY}`, () => {
-          cy.assertChangeAnswersPageUrl({ referenceNumber, route: TYPE_OF_POLICY, fieldId });
-        });
-
-        it('should render the new answer', () => {
-          const { day, month, year } = newStartDate;
-
-          fieldVariables.startDate.newValue = formatDate(createTimestampFromNumbers(day, month, year));
-          cy.checkChangeAnswerRendered({ fieldVariables: fieldVariables.startDate });
-        });
-      });
-    });
-
-    describe(CONTRACT_COMPLETION_DATE, () => {
-      const fieldId = CONTRACT_COMPLETION_DATE;
-      let fieldVariables = getFieldVariables(fieldId, referenceNumber);
-
-      const newAnswer = {
-        ...application.POLICY[fieldId],
-        year: application.POLICY[fieldId].year + 2,
-      };
-
-      describe('when clicking the `change` link', () => {
-        beforeEach(() => {
-          cy.navigateToUrl(url);
-        });
-
-        it(`should redirect to ${SINGLE_CONTRACT_POLICY_CHECK_AND_CHANGE}`, () => {
-          cy.navigateToUrl(url);
-          fieldVariables = getFieldVariables(fieldId, referenceNumber);
 
           cy.checkChangeLinkUrl(fieldVariables, referenceNumber);
         });
@@ -177,27 +118,58 @@ context('Insurance - Change your answers - Policy - Single contract policy - Sum
       });
     });
 
-    describe(TOTAL_CONTRACT_VALUE, () => {
-      const fieldId = TOTAL_CONTRACT_VALUE;
+    describe(TOTAL_MONTHS_OF_COVER, () => {
+      const fieldId = TOTAL_MONTHS_OF_COVER;
 
-      const fieldVariables = {
-        route: SINGLE_CONTRACT_POLICY_TOTAL_CONTRACT_VALUE_CHECK_AND_CHANGE,
-        newValueInput: application.POLICY[fieldId] - 500,
-        fieldId,
-        referenceNumber,
-        summaryList,
-        changeLink: summaryList.field(fieldId).changeLink,
-      };
+      const fieldVariables = getFieldVariables(fieldId, referenceNumber);
 
       describe('when clicking the `change` link', () => {
         beforeEach(() => {
           cy.navigateToUrl(url);
+        });
+
+        it(`should redirect to ${MULTIPLE_CONTRACT_POLICY_CHECK_AND_CHANGE}`, () => {
+          cy.navigateToUrl(url);
 
           cy.checkChangeLinkUrl(fieldVariables, referenceNumber);
         });
+      });
 
-        it(`should redirect to ${SINGLE_CONTRACT_POLICY_TOTAL_CONTRACT_VALUE_CHECK_AND_CHANGE}`, () => {
-          cy.assertChangeAnswersPageUrl({ referenceNumber, route: fieldVariables.route, fieldId });
+      describe('form submission with a new answer', () => {
+        beforeEach(() => {
+          cy.navigateToUrl(url);
+
+          summaryList.field(fieldId).changeLink().click();
+
+          fieldVariables.newValueInput = String(Number(application.POLICY[fieldId]) + 1);
+          cy.changeAnswerField(fieldVariables, field(fieldId).input());
+        });
+
+        it(`should redirect to ${TYPE_OF_POLICY}`, () => {
+          cy.assertChangeAnswersPageUrl({ referenceNumber, route: TYPE_OF_POLICY, fieldId });
+        });
+
+        it('should render the new answer', () => {
+          fieldVariables.newValue = `${fieldVariables.newValueInput} months`;
+          cy.checkChangeAnswerRendered({ fieldVariables });
+        });
+      });
+    });
+
+    describe(TOTAL_SALES_TO_BUYER, () => {
+      const fieldId = TOTAL_SALES_TO_BUYER;
+
+      const fieldVariables = getFieldVariables(fieldId, referenceNumber, exportValueRoute);
+
+      describe('when clicking the `change` link', () => {
+        beforeEach(() => {
+          cy.navigateToUrl(url);
+        });
+
+        it(`should redirect to ${exportValueRoute}`, () => {
+          cy.navigateToUrl(url);
+
+          cy.checkChangeLinkUrl(fieldVariables, referenceNumber);
         });
       });
 
@@ -222,8 +194,47 @@ context('Insurance - Change your answers - Policy - Single contract policy - Sum
       });
     });
 
+    describe(MAXIMUM_BUYER_WILL_OWE, () => {
+      const fieldId = MAXIMUM_BUYER_WILL_OWE;
+
+      const fieldVariables = getFieldVariables(fieldId, referenceNumber, exportValueRoute);
+
+      describe('when clicking the `change` link', () => {
+        beforeEach(() => {
+          cy.navigateToUrl(url);
+        });
+
+        it(`should redirect to ${MULTIPLE_CONTRACT_POLICY_CHECK_AND_CHANGE}`, () => {
+          cy.navigateToUrl(url);
+
+          cy.checkChangeLinkUrl(fieldVariables, referenceNumber);
+        });
+      });
+
+      describe('form submission with a new answer', () => {
+        beforeEach(() => {
+          cy.navigateToUrl(url);
+
+          summaryList.field(fieldId).changeLink().click();
+
+          fieldVariables.newValueInput = Number(application.POLICY[fieldId]) + 1000;
+          cy.changeAnswerField(fieldVariables, multipleContractPolicyExportValuePage[fieldId].input());
+        });
+
+        it(`should redirect to ${TYPE_OF_POLICY}`, () => {
+          cy.assertChangeAnswersPageUrl({ referenceNumber, route: TYPE_OF_POLICY, fieldId });
+        });
+
+        it('should render the new answer', () => {
+          fieldVariables.newValue = formatCurrency(fieldVariables.newValueInput);
+          cy.checkChangeAnswerRendered({ fieldVariables });
+        });
+      });
+    });
+
     describe(CURRENCY_CODE, () => {
       const fieldId = CURRENCY_CODE;
+
       const fieldVariables = {
         ...getFieldVariables(fieldId, referenceNumber),
         newValueInput: USD.isoCode,
@@ -235,7 +246,7 @@ context('Insurance - Change your answers - Policy - Single contract policy - Sum
           cy.navigateToUrl(url);
         });
 
-        it(`should redirect to ${SINGLE_CONTRACT_POLICY_CHECK_AND_CHANGE}`, () => {
+        it(`should redirect to ${MULTIPLE_CONTRACT_POLICY_CHECK_AND_CHANGE}`, () => {
           cy.navigateToUrl(url);
 
           cy.checkChangeLinkUrl(fieldVariables, referenceNumber);
@@ -255,10 +266,8 @@ context('Insurance - Change your answers - Policy - Single contract policy - Sum
           cy.assertChangeAnswersPageUrl({ referenceNumber, route: TYPE_OF_POLICY, fieldId });
         });
 
-        it('should render the new answer and retain a `completed` status tag', () => {
+        it('should render the new answer', () => {
           cy.checkChangeAnswerRendered({ fieldVariables });
-
-          cy.checkTaskStatusCompleted(status);
         });
       });
     });
