+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/policy/change-your-answers/multiple-policy/change-your-answers-policy-type-change-multiple-to-single-policy-type-summary-list.spec.js
@@ -1,12 +1,12 @@
-import partials from '../../../../../../../partials';
-import { FIELD_VALUES } from '../../../../../../../constants';
-import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
-import { INSURANCE_FIELD_IDS } from '../../../../../../../constants/field-ids/insurance';
-import { summaryList } from '../../../../../../../pages/shared';
-import { typeOfPolicyPage } from '../../../../../../../pages/insurance/policy';
-import { createTimestampFromNumbers, formatDate } from '../../../../../../../helpers/date';
-import formatCurrency from '../../../../../../../helpers/format-currency';
-import application from '../../../../../../../fixtures/application';
+import partials from '../../../../../../../../partials';
+import { FIELD_VALUES } from '../../../../../../../../constants';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+import { INSURANCE_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance';
+import { summaryList } from '../../../../../../../../pages/shared';
+import { typeOfPolicyPage } from '../../../../../../../../pages/insurance/policy';
+import { createTimestampFromNumbers, formatDate } from '../../../../../../../../helpers/date';
+import formatCurrency from '../../../../../../../../helpers/format-currency';
+import application from '../../../../../../../../fixtures/application';
 
 const {
   ROOT,
