+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/policy/change-your-answers/name-on-policy/change-your-answers-policy-change-same-name-to-different-name-on-policy.spec.js
@@ -1,8 +1,8 @@
-import partials from '../../../../../../../partials';
-import { INSURANCE_FIELD_IDS } from '../../../../../../../constants/field-ids/insurance';
-import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
-import { summaryList } from '../../../../../../../pages/shared';
-import application from '../../../../../../../fixtures/application';
+import partials from '../../../../../../../../partials';
+import { INSURANCE_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+import { summaryList } from '../../../../../../../../pages/shared';
+import application from '../../../../../../../../fixtures/application';
 
 const {
   ROOT: INSURANCE_ROOT,
