+++ e2e-tests/insurance/cypress/e2e/journeys/check-your-answers/policy/summary-list/single-policy/check-your-answers-policy-single-with-broker-summary-list.spec.js
@@ -1,7 +1,7 @@
-import partials from '../../../../../../../partials';
-import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
-import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../constants/field-ids/insurance/policy';
-import checkSummaryList from '../../../../../../../commands/insurance/check-policy-summary-list';
+import partials from '../../../../../../../../partials';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/policy';
+import checkSummaryList from '../../../../../../../../commands/insurance/check-policy-summary-list';
 
 const {
   ROOT: INSURANCE_ROOT,
